examples/openstack-app/requirements.txt
keystoneauth1==2.15.0 # Apache-2.0
python-novaclient==6.0.0
python-keystoneclient==3.6.0 # Apache-2.0
